+++ R/db_utils_not_exported.R
@@ -1251,3 +1251,79 @@ detect_crs <- function(x) {
   return(crs_x)
 
 }
+
+
+
+
+
+#' Helper for geometry validation functions (e.g. ST_IsValid)
+#'
+#' @template x
+#' @template conn_null
+#' @template crs
+#' @template quiet
+#' @param fun The duckdb function to use
+#' 
+#' @return A logical vector
+#' @keywords internal
+#' @noRd
+geometry_validation_helper <- function(
+  x,
+  conn = NULL,
+  crs = NULL,
+  crs_column = "crs_duckspatial",
+  quiet = FALSE,
+  fun) {
+  
+  deprecate_crs(crs_column, crs)
+
+  ## 0. Handle errors
+  assert_xy(x, "x")
+  assert_conn_character(conn, x)
+  assert_logic(quiet, "quiet")
+
+
+  # 1. Manage connection to DB
+
+  ## 1.1. Pre-extract attributes (CRS and geometry column name)
+  ## this step should be before normalize_spatial_input()
+  crs_x    <- ddbs_crs(x, conn)
+  sf_col_x <- attr(x, "sf_column")
+
+  ## 1.2. Normalize inputs: coerce tbl_duckdb_connection to duckspatial_df, 
+  ## validate character table names
+  x <- normalize_spatial_input(x, conn)
+
+
+  # 2. Manage connection to DB
+
+  ## 2.1. Resolve connections and handle imports
+  resolve_conn <- resolve_spatial_connections(x, y = NULL, conn = conn)
+  target_conn  <- resolve_conn$conn
+  x            <- resolve_conn$x
+  ## register cleanup of the connection
+  on.exit(resolve_conn$cleanup(), add = TRUE)
+
+  ## 2.2. Get query list of table names
+  x_list <- get_query_list(x, target_conn)
+  on.exit(x_list$cleanup(), add = TRUE)
+
+
+  # 3. Make the query
+
+  ## 3.1. Prepare parameters
+  x_geom <- sf_col_x %||% get_geom_name(target_conn, x_list$query_name)
+  assert_geometry_column(x_geom, x_list)
+
+  ## 3.2. Prepare the query
+  tmp.query <- glue::glue("
+    SELECT {fun}({x_geom}) as res,
+    FROM {x_list$query_name}
+  ")
+
+  ## 3.3. Get results
+  data_vec <- DBI::dbGetQuery(target_conn, tmp.query)
+  feedback_query(quiet)
+  return(data_vec[, 1])
+  
+}